+++ src/enums/timeline.ts
@@ -1,0 +1,19 @@
+export enum ActivityLayerFilterField {
+  'Type' = 'Type',
+  'Name' = 'Name',
+  'Subsystem' = 'Subsystem',
+  'Tags' = 'Tags',
+  'Parameter' = 'Parameter',
+  'SchedulingGoalId' = 'Scheduling Goal Id',
+}
+
+export enum FilterOperator {
+  'equals' = 'equals',
+  'does_not_equal' = 'does not equal',
+  'includes' = 'includes',
+  'does_not_include' = 'does not include',
+  'is_greater_than' = 'is greater than',
+  'is_less_than' = 'is less than',
+  'is_within' = 'is within',
+  'is_not_within' = 'is not within',
+}
